refactor(artifact-services-dotnet-asp): split artifact detail types
