+++ modules/artifact-services-dotnet-asp/src/main/kotlin/io/github/lmliam/microsmith/artifact/services/dotnet/asp/DotnetAspHeaderFieldArtifact.kt
@@ -1,0 +1,6 @@
+package io.github.lmliam.microsmith.artifact.services.dotnet.asp
+
+data class DotnetAspHeaderFieldArtifact(
+    val name: String,
+    val headerName: String,
+)
