+++ modules/artifact-services-dotnet-asp/src/main/kotlin/io/github/lmliam/microsmith/artifact/services/dotnet/asp/DotnetAspHeadersBindingArtifact.kt
@@ -1,0 +1,8 @@
+package io.github.lmliam.microsmith.artifact.services.dotnet.asp
+
+data class DotnetAspHeadersBindingArtifact(
+    val typeName: String,
+    val name: String,
+    val headers: List<DotnetAspHeaderFieldArtifact>,
+    val origins: Set<String>,
+)
