+++ modules/artifact-services-dotnet-asp/src/main/kotlin/io/github/lmliam/microsmith/artifact/services/dotnet/asp/DotnetAspModelArtifact.kt
@@ -1,0 +1,10 @@
+package io.github.lmliam.microsmith.artifact.services.dotnet.asp
+
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.model.DotnetModel
+
+data class DotnetAspModelArtifact(
+    val typeName: String,
+    val locality: DotnetAspModelLocality,
+    val model: DotnetModel,
+    val origins: Set<String>,
+)
