+++ modules/artifact-services-dotnet-asp/src/main/kotlin/io/github/lmliam/microsmith/artifact/services/dotnet/asp/DotnetAspModelLocality.kt
@@ -1,0 +1,6 @@
+package io.github.lmliam.microsmith.artifact.services.dotnet.asp
+
+enum class DotnetAspModelLocality {
+    SHARED,
+    INLINE,
+}
