+++ modules/artifact-services-dotnet-asp/src/main/kotlin/io/github/lmliam/microsmith/artifact/services/dotnet/asp/DotnetAspRequestBindingArtifact.kt
@@ -1,0 +1,10 @@
+package io.github.lmliam.microsmith.artifact.services.dotnet.asp
+
+import io.github.lmliam.microsmith.dsl.services.dotnet.core.model.DotnetFieldType
+
+data class DotnetAspRequestBindingArtifact(
+    val typeName: String,
+    val name: String,
+    val fields: List<DotnetAspRequestFieldArtifact>,
+    val origins: Set<String>,
+)
