+++ modules/artifact-services-dotnet-asp/src/main/kotlin/io/github/lmliam/microsmith/artifact/services/dotnet/asp/DotnetAspResponseArtifact.kt
@@ -1,0 +1,8 @@
+package io.github.lmliam.microsmith.artifact.services.dotnet.asp
+
+data class DotnetAspResponseArtifact(
+    val statusCode: Int,
+    val model: DotnetAspModelArtifact,
+    val headers: List<DotnetAspResponseHeaderArtifact>,
+    val origins: Set<String>,
+)
