+++ modules/artifact-services-dotnet-asp/src/main/kotlin/io/github/lmliam/microsmith/artifact/services/dotnet/asp/DotnetAspResponseHeaderArtifact.kt
@@ -1,0 +1,3 @@
+package io.github.lmliam.microsmith.artifact.services.dotnet.asp
+
+data class DotnetAspResponseHeaderArtifact(val name: String)
